+++ lib/screens/layouts/desktop/viewport.dart
@@ -20,7 +20,7 @@
 import 'dart:async';
 
 import 'package:bluecherry_client/models/device.dart';
-import 'package:bluecherry_client/providers/desktop_view_provider.dart';
+import 'package:bluecherry_client/providers/layouts_provider.dart';
 import 'package:bluecherry_client/providers/settings_provider.dart';
 import 'package:bluecherry_client/screens/layouts/desktop/multicast_view.dart';
 import 'package:bluecherry_client/screens/layouts/desktop/stream_data.dart';
