+++ lib/screens/multi_window/window.dart
@@ -18,7 +18,7 @@
  */
 
 import 'package:bluecherry_client/main.dart';
-import 'package:bluecherry_client/providers/desktop_view_provider.dart';
+import 'package:bluecherry_client/providers/layouts_provider.dart';
 import 'package:bluecherry_client/providers/home_provider.dart';
 import 'package:bluecherry_client/providers/settings_provider.dart';
 import 'package:bluecherry_client/utils/theme.dart';
